Move `lib/api/form` to `app/component/form`
