+++ frontend/app/component/form/data_entry/check_and_save/CheckAndSaveForm.test.tsx
@@ -3,12 +3,13 @@ import * as router from "react-router";
 import { userEvent } from "@testing-library/user-event";
 import { beforeEach, describe, expect, test, vi } from "vitest";
 
-import { CheckAndSaveForm } from "app/component/form/data_entry/check_and_save/CheckAndSaveForm";
-import { defaultFormState, emptyDataEntryRequest, errorWarningMocks } from "app/component/form/testHelperFunctions.ts";
-
-import { ElectionProvider, FormState, PollingStationFormController, PollingStationResults } from "@kiesraad/api";
+import { ElectionProvider, PollingStationResults } from "@kiesraad/api";
 import { electionDetailsMockResponse, electionMockData } from "@kiesraad/api-mocks";
 import { overrideOnce, render, screen, server, within } from "@kiesraad/test";
+
+import { defaultFormState, emptyDataEntryRequest, errorWarningMocks } from "../../testHelperFunctions";
+import { FormState, PollingStationFormController } from "../PollingStationFormController";
+import { CheckAndSaveForm } from "./CheckAndSaveForm";
 
 const mockNavigate = vi.fn();
