+++ frontend/app/component/form/data_entry/differences/DifferencesForm.test.tsx
@@ -9,14 +9,11 @@ import {
   expectFieldsToNotHaveIcon,
 } from "app/component/form/testHelperFunctions.ts";
 
-import {
-  POLLING_STATION_DATA_ENTRY_SAVE_REQUEST_BODY,
-  PollingStationFormController,
-  PollingStationResults,
-} from "@kiesraad/api";
+import { POLLING_STATION_DATA_ENTRY_SAVE_REQUEST_BODY, PollingStationResults } from "@kiesraad/api";
 import { electionMockData, pollingStationMockData } from "@kiesraad/api-mocks";
 import { getUrlMethodAndBody, overrideOnce, render, screen, userTypeInputs } from "@kiesraad/test";
 
+import { PollingStationFormController } from "../PollingStationFormController";
 import { DifferencesForm } from "./DifferencesForm";
 
 function renderForm(defaultValues: Partial<PollingStationResults> = {}) {
