+++ frontend/app/component/form/data_entry/differences/useDifferences.ts
@@ -1,6 +1,8 @@
 import * as React from "react";
 
-import { PollingStationResults, usePollingStationFormController } from "@kiesraad/api";
+import { PollingStationResults } from "@kiesraad/api";
+
+import { usePollingStationFormController } from "../usePollingStationFormController";
 
 export type DifferencesValues = Pick<PollingStationResults, "differences_counts">;
