+++ frontend/app/component/form/data_entry/recounted/RecountedForm.test.tsx
@@ -3,10 +3,11 @@ import { describe, expect, test, vi } from "vitest";
 
 import { emptyDataEntryRequest } from "app/component/form/testHelperFunctions.ts";
 
-import { POLLING_STATION_DATA_ENTRY_SAVE_REQUEST_BODY, PollingStationFormController } from "@kiesraad/api";
+import { POLLING_STATION_DATA_ENTRY_SAVE_REQUEST_BODY } from "@kiesraad/api";
 import { electionMockData } from "@kiesraad/api-mocks";
 import { getUrlMethodAndBody, overrideOnce, render, screen } from "@kiesraad/test";
 
+import { PollingStationFormController } from "../PollingStationFormController";
 import { RecountedForm } from "./RecountedForm";
 
 const Component = (
