+++ frontend/app/module/data_entry/polling_station/PollingStationLayout.test.tsx
@@ -4,15 +4,11 @@ import { beforeEach, describe, expect, test, vi } from "vitest";
 
 import { PollingStationLayout } from "app/module/data_entry";
 
-import {
-  Election,
-  ElectionListProvider,
-  ElectionProvider,
-  ElectionStatusProvider,
-  PollingStationFormController,
-} from "@kiesraad/api";
+import { Election, ElectionListProvider, ElectionProvider, ElectionStatusProvider } from "@kiesraad/api";
 import { electionDetailsMockResponse, pollingStationMockData } from "@kiesraad/api-mocks";
 import { overrideOnce, render, screen, within } from "@kiesraad/test";
+
+import { PollingStationFormController } from "../../../component/form/data_entry/PollingStationFormController";
 
 describe("PollingStationLayout", () => {
   const election = electionDetailsMockResponse.election as Required<Election>;
