--- frontend/lib/api/form/index.ts
@@ -1,1 +1,0 @@
-export * from "./pollingstation";
