--- frontend/lib/api/form/pollingstation/index.ts
@@ -1,7 +1,0 @@
-export * from "./PollingStationFormController";
-export * from "./pollingStationUtils";
-export * from "./useDifferences";
-export * from "./usePoliticalGroups";
-export * from "./usePollingStationFormController";
-export * from "./useRecounted";
-export * from "./useVotersAndVotes";
