+++ frontend/lib/api/index.ts
@@ -1,5 +1,4 @@
 export * from "./election";
-export * from "./form";
 export * from "./gen/openapi";
 export * from "./polling_station";
 export type * from "./api.types";
